rsibot/requirements.txt
python-binance
TA-Lib
numpy
websocket_client
python-dotenv
pymongo